chore: Refine the installation script

+++ install.sh
@@ -157,7 +157,10 @@ function Install_Docker(){
         if [[ $major_version -lt 20 ]]; then
             log "$TXT_LOW_DOCKER_VERSION"
         fi
-        configure_accelerator
+
+        if [[ $(curl -s ipinfo.io/country) == "CN" ]]; then
+            configure_accelerator
+        fi
     else
         log "$TXT_DOCKER_INSTALL_ONLINE"
 
@@ -283,7 +286,13 @@ function Install_Compose(){
 		if [ "$arch" == 'armv7l' ]; then
 			arch='armv7'
 		fi
-		curl -L https://resource.fit2cloud.com/docker/compose/releases/download/v2.26.1/docker-compose-$(uname -s | tr A-Z a-z)-"$arch" -o /usr/local/bin/docker-compose 2>&1 | tee -a "${CURRENT_DIR}"/install.log
+
+        if [[ $(curl -s ipinfo.io/country) == "CN" ]]; then
+		    curl -L https://resource.fit2cloud.com/docker/compose/releases/download/v2.26.1/docker-compose-$(uname -s | tr A-Z a-z)-"$arch" -o /usr/local/bin/docker-compose 2>&1 | tee -a "${CURRENT_DIR}"/install.log
+        else
+            curl -L https://github.com/docker/compose/releases/download/v2.26.1/docker-compose-$(uname -s | tr A-Z a-z)-"$arch" -o /usr/local/bin/docker-compose 2>&1 | tee -a "${CURRENT_DIR}"/install.log
+        fi
+
         if [[ ! -f /usr/local/bin/docker-compose ]];then
             log "$TXT_DOCKER_COMPOSE_DOWNLOAD_FAIL"
             exit 1
